refactor: 폴더 구조 수정
[#480]
